requirements.txt
coverage==4.5.1
Django==2.1.2
django-cors-headers==2.4.0
django-filter==2.0.0
djangorestframework==3.8.2
flake8==3.6.0
ipdb==0.13.9
isort==4.3.4
gunicorn==19.9.0
Markdown==3.0.1
psycopg2-binary==2.7.5
pytz==2018.5
